LPS-144455 Separate exception mapper to another class.
When jaxrs compares exception mappers to find the most specific one, it uses the first interface https://github.com/apache/cxf/blob/88ba228e96a81aeded38228ef94914612a7968be/rt/frontend/jaxrs/src/main/java/org/apache/cxf/jaxrs/provider/ProviderFactory.java#L1157. Having more than one result in this comparison not being done correctly, this is why this exception mapper was never used
